Sort out messages
